refactor(sync): extract sync_row_meta accessors + LocalMeta into apply/meta.rs
